08-BusinessIntelligence/Contract-Intelligence-Platform-Agent/requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
langchain-mcp-adapters>=0.1.0
mcp>=1.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
streamlit>=1.35.0
nest-asyncio>=1.6.0
uvicorn>=0.30.0
httpx>=0.27.0
redis>=5.0.0
fastapi>=0.115.0
PyJWT>=2.8.0
bcrypt>=4.1.0
pandas>=2.0.0
numpy>=1.26.0
